debug issue with get_cards query by changing price typedef to Float

+++ backend/schema/typedefs.js
@@ -36,7 +36,7 @@ type Card {
   image: String!
   cardId: String!
   setId: String!
-  price: Int
+  price: Float
 }
 
 type Query {
